# backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: College Lost & Found API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register with college email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: OTP sent to email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid email or already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login and request OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OTP sent to email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /items/{id}:
    get:
      operationId: getItem
      tags: [items]
      summary: Get a single item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Item details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an item (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /claims:
    post:
      operationId: createClaim
      tags: [claims]
      summary: Initiate a claim on a found item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClaimRequest"
      responses:
        "201":
          description: Claim created and OTP generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResponse"
        "400":
          description: Cannot claim own item or item already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /claims/{id}/verify:
    post:
      operationId: verifyClaim
      tags: [claims]
      summary: Finder enters OTP to confirm item returned
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyClaimRequest"
      responses:
        "200":
          description: Item marked as returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, faculty, admin]
      required: [name, email, password, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
...
    User:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        role:
          type: string
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, isAdmin, createdAt]
    Item:
      type: object
      properties:
        id:
          type: number
        type:
          type: string
          enum: [lost, found]
        title:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        imageUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, claimed, returned]
        reportedBy:
          type: number
        reporterName:
          type: string
        reporterEmail:
          type: string
        createdAt:
          type: string
        claimCount:
          type: number
      required: [id, type, title, description, category, location, status, reportedBy, reporterName, reporterEmail, createdAt, claimCount]
...
    CreateClaimRequest:
      type: object
      properties:
        itemId:
          type: number
        message:
          type: string
      required: [itemId, message]
    ClaimResponse:
      type: object
      properties:
        id:
          type: number
        itemId:
          type: number
        claimerId:
          type: number
        status:
          type: string
        otp:
          type: string
          description: OTP for the claimant to share with the finder
        message:
          type: string
        createdAt:
          type: string
      required: [id, itemId, claimerId, status, otp, message, createdAt]
    VerifyClaimRequest:
      type: object
      properties:
        otp:
          type: string
      required: [otp]
...